requirements.txt
python-dotenv==1.0.0
streamlit==1.18.1
